feat: makes command palette a button and usable as a component
