+++ frontend/src/lib/components/header.svelte
@@ -4,6 +4,8 @@
 	import NavigationItem from '$lib/components/header-item.svelte';
 	import { Mountain, MoreHorizontal, X, Command } from 'lucide-svelte';
 	import { Button } from '$lib/components/ui/button';
+	import CommandItem from '$lib/components/command-item.svelte';
+	import { commandPalette } from '$lib/stores';
 
 	const navItems: NavItem[] = [
 		{
@@ -22,6 +24,10 @@
 
 	let showMenu = false;
 
+	function toggleCommand() {
+		$commandPalette = !$commandPalette;
+	}
+
 	function toggleNavbar() {
 		showMenu = !showMenu;
 	}
@@ -31,19 +37,21 @@
 	<div class="flex items-center gap-2">
 		<a href="/" class="flex items-center gap-2">
 			<Mountain class="size-6 md:size-8" />
-			<h1 class="text-xl md:text-2xl font-semibold">Iceberg</h1>
+			<h1 class="text-xl font-semibold md:text-2xl">Iceberg</h1>
 		</a>
-		<div
-			class="ml-2 hidden lg:flex items-center p-2 px-4 bg-secondary rounded-md text-sm font-medium"
+		<Button
+			class="items-center hidden p-2 px-4 ml-2 text-sm font-medium rounded-md lg:flex"
+			type="button"
+			on:click={toggleCommand}
 		>
 			<div class="flex items-center">
-				<Command class="h-4 w-4" />
+				<Command class="w-4 h-4" />
 				<span>K</span>
 			</div>
 			<span class="ml-2">to open command palette</span>
-		</div>
+		</Button>
 	</div>
-	<nav class="items-center gap-6 tracking-wider hidden md:flex">
+	<nav class="items-center hidden gap-6 tracking-wider md:flex">
 		<div class="flex items-center gap-3">
 			{#each navItems as navItem}
 				<NavigationItem {navItem} />
@@ -54,7 +62,7 @@
 	<nav class="flex items-center gap-2 tracking-wider md:hidden">
 		<ThemeSwitcher />
 		<Button on:click={toggleNavbar} type="button" size="sm" class="max-w-max">
-			<MoreHorizontal class="h-4 w-4" />
+			<MoreHorizontal class="w-4 h-4" />
 		</Button>
 	</nav>
 </header>
@@ -66,12 +74,12 @@
 	class:h-[100dvh]={showMenu}
 	class="h-0 w-screen fixed z-[99] top-0 left-0 bg-background overflow-x-hidden flex flex-col items-center md:hidden"
 >
-	<div class="flex p-8 w-full items-end justify-end transition-all ease-in-out duration-300">
+	<div class="flex items-end justify-end w-full p-8 transition-all duration-300 ease-in-out">
 		<Button on:click={toggleNavbar} type="button" size="sm" class="max-w-max">
-			<X class="h-4 w-4" />
+			<X class="w-4 h-4" />
 		</Button>
 	</div>
-	<div class="flex flex-col items-center justify-center gap-6 p-8 w-full">
+	<div class="flex flex-col items-center justify-center w-full gap-6 p-8">
 		{#each navItems as navItem}
 			<button on:click={toggleNavbar}>
 				<NavigationItem {navItem} />
@@ -79,6 +87,8 @@
 		{/each}
 	</div>
 </div>
+
+<CommandItem bind:open={$commandPalette}/>
 
 <style>
 	#mobilenav {
