+++ frontend/src/lib/stores.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store'
+
+export const commandPalette = writable(false)
